backend/requirements.txt
django
djangorestframework
cloudinary
python-decouple
#instalar cors
django-cors-headers
djangorestframework-simplejwt